zkit lpc1768: Rename files for better conformance to naming conventions
