+++ nuttx/configs/zkit-arm-1769/src/lpc17_can.c
@@ -1,12 +1,12 @@
 /************************************************************************************
- * configs/zkit-arm-1769/src/up_boot.c
+ * configs/zkit-arm-1769/src/lpc17_can.c
  *
  *   Copyright (C) 2013 Zilogic Systems. All rights reserved.
- *   Author: BabuSubashChandar <code@zilogic.com>
+ *   Author: Raashid Muhammed <code@zilogic.com>
  *
- * Based on configs/lpcxpresso-lpc1768/src/up_boot.c
+ *   Based on configs/olimex-lpc1766stk/src/lpc17_can.c
  *
- *   Copyright (C) 2011 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -44,19 +44,42 @@
 
 #include <nuttx/config.h>
 
+#include <errno.h>
 #include <debug.h>
 
-#include <nuttx/board.h>
+#include <nuttx/can.h>
 #include <arch/board/board.h>
 
+#include "chip.h"
 #include "up_arch.h"
-#include "up_internal.h"
 
+#include "lpc17_can.h"
 #include "zkitarm_internal.h"
+
+#if defined(CONFIG_CAN) && (defined(CONFIG_LPC17_CAN1) || defined(CONFIG_LPC17_CAN2))
 
 /************************************************************************************
  * Pre-processor Definitions
  ************************************************************************************/
+/* Configuration ********************************************************************/
+#define CAN_PORT1 1
+#define CAN_PORT2 2
+
+
+/* Debug ***************************************************************************/
+/* Non-standard debug that may be enabled just for testing CAN */
+
+#ifdef CONFIG_DEBUG_CAN
+#  define candbg    dbg
+#  define canvdbg   vdbg
+#  define canlldbg  lldbg
+#  define canllvdbg llvdbg
+#else
+#  define candbg(x...)
+#  define canvdbg(x...)
+#  define canlldbg(x...)
+#  define canllvdbg(x...)
+#endif
 
 /************************************************************************************
  * Private Functions
@@ -67,31 +90,70 @@
  ************************************************************************************/
 
 /************************************************************************************
- * Name: lpc17_boardinitialize
+ * Name: can_devinit
  *
  * Description:
- *   All LPC17xx architectures must provide the following entry point.  This entry point
- *   is called early in the intitialization -- after all memory has been configured
- *   and mapped but before any devices have been initialized.
+ *   All LPC17 architectures must provide the following interface to work with
+ *   examples/can.
  *
  ************************************************************************************/
 
-void lpc17_boardinitialize(void)
+int can_devinit(void)
 {
-  /* Configure SSP chip selects if 1) at least one SSP is enabled, and 2) the weak
-   * function zkit_sspinitialize() has been brought into the link.
-   */
+  static bool initialized = false;
+  struct can_dev_s *can;
+  int ret;
 
-#if defined(CONFIG_LPC17_SSP0) || defined(CONFIG_LPC17_SSP1)
-  if (zkit_sspinitialize)
+  /* Check if we have already initialized */
+
+  if (!initialized)
     {
-      zkit_sspinitialize();
+#ifdef CONFIG_LPC17_CAN1
+      /* Call lpc17_caninitialize() to get an instance of the CAN1 interface */
+
+      can = lpc17_caninitialize(CAN_PORT1);
+      if (can == NULL)
+        {
+          candbg("ERROR:  Failed to get CAN1 interface\n");
+          return -ENODEV;
+        }
+
+      /* Register the CAN1 driver at "/dev/can0" */
+
+      ret = can_register("/dev/can0", can);
+      if (ret < 0)
+        {
+          candbg("ERROR: CAN1 register failed: %d\n", ret);
+          return ret;
+        }
+#endif
+
+#ifdef CONFIG_LPC17_CAN2
+      /* Call lpc17_caninitialize() to get an instance of the CAN2 interface */
+
+      can = lpc17_caninitialize(CAN_PORT2);
+      if (can == NULL)
+        {
+          candbg("ERROR:  Failed to get CAN2 interface\n");
+          return -ENODEV;
+        }
+
+      /* Register the CAN2 driver at "/dev/can1" */
+
+      ret = can_register("/dev/can1", can);
+      if (ret < 0)
+        {
+          candbg("ERROR: CAN2 register failed: %d\n", ret);
+          return ret;
+        }
+#endif
+
+      /* Now we are initialized */
+
+      initialized = true;
     }
-#endif
 
-  /* Configure on-board LEDs if LED support has been selected. */
-
-#ifdef CONFIG_ARCH_LEDS
-  board_led_initialize();
-#endif
+  return OK;
 }
+
+#endif /* CONFIG_CAN && (CONFIG_LPC17_CAN1 || CONFIG_LPC17_CAN2) */
